Refactor: Move BulkIngestService to domain package and move methods to ImageService and StorageService
